1、机房删除时，如果包含机柜则不允许删除。

+++ src/main/java/com/asiainfo/aiom/view/system/location/LocationDeleteAction.java
@@ -9,7 +9,13 @@
  */
 package com.asiainfo.aiom.view.system.location;
 
+import java.util.List;
+
+import com.asiainfo.gim.client.site.api.CabinetApi;
 import com.asiainfo.gim.client.site.api.LocationApi;
+import com.asiainfo.gim.client.site.domain.Cabinet;
+import com.asiainfo.gim.client.site.domain.query.CabinetQueryCondition;
+import com.asiainfo.support.struts2.ResultBean;
 import com.asiainfo.support.struts2.ServletAwareActionSupport;
 
 /**
@@ -21,6 +27,10 @@ public class LocationDeleteAction extends ServletAwareActionSupport
 	private static final long serialVersionUID = -2816604616177215575L;
 	
 	private LocationApi locationApi;
+	
+	private CabinetApi cabinetApi;
+	
+	private ResultBean resultBean;
 	
 	private Integer id;
 
@@ -39,9 +49,36 @@ public class LocationDeleteAction extends ServletAwareActionSupport
 		this.locationApi = locationApi;
 	}
 	
+	public void setCabinetApi(CabinetApi cabinetApi)
+	{
+		this.cabinetApi = cabinetApi;
+	}
+
+	public ResultBean getResultBean()
+	{
+		return resultBean;
+	}
+
+	public void setResultBean(ResultBean resultBean)
+	{
+		this.resultBean = resultBean;
+	}
+
 	public String execute()
 	{
-		locationApi.deleteLocation(id);
+		CabinetQueryCondition cabinetQueryCondition = new CabinetQueryCondition();
+		cabinetQueryCondition.setLocationId(id);
+		List<Cabinet> cabinets = cabinetApi.listCabinets(cabinetQueryCondition);
+		
+		if (cabinets != null && cabinets.size() > 0)
+		{
+			resultBean = new ResultBean(false, "此机房下存在机柜，无法删除！");
+		}
+		else
+		{
+			locationApi.deleteLocation(id);
+			resultBean = new ResultBean(true, null);
+		}
 		return SUCCESS;
 	}
 }
